backend/requirements.txt
# FastAPI and server
fastapi==0.128.0
uvicorn[standard]==0.24.0
python-multipart>=0.0.6

# HTTP client for AIDP integration
httpx>=0.25.2

# Image processing
Pillow>=10.1.0

# Memory profiling
psutil>=5.9.0

# GLTF file validation
pygltflib>=1.16.1

# YAML parsing for scene presets
PyYAML>=6.0

# Template rendering for Blender scripts
Jinja2>=3.1.0

# Job scheduling and cleanup
APScheduler>=3.10.4

# Testing
pytest==9.0.2
pytest-asyncio>=0.21.1

# Code formatting and linting
black>=24.1.1
isort>=5.13.2

# Environment configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0
